+++ time/src/client/mod.rs
@@ -9,26 +9,31 @@
 
 #[cfg(feature = "tcp-client")]
 mod tcp;
-#[cfg(feature = "tcp-client")]
-pub use tcp::*;
 
+use async_trait::async_trait;
 use log::{info, trace};
 use std::io;
 
+use crate::{RequestWriter, ResponseReader};
+
 use super::{Request, Response, Timer};
+
+#[cfg(feature = "tcp-client")]
+pub use self::tcp::*;
 
 /// The client trait.
 ///
 /// Clients must implement this trait. Only the [`Client::send`]
 /// function needs to be implemented: it should describe how to
 /// connect and send requests to the server.
-pub trait Client {
-    fn send(&self, req: Request) -> io::Result<Response>;
+#[async_trait]
+pub trait Client: Send + Sync {
+    async fn send(&self, req: Request) -> io::Result<Response>;
 
-    fn start(&self) -> io::Result<()> {
+    async fn start(&self) -> io::Result<()> {
         info!("sending request to start timer");
 
-        match self.send(Request::Start) {
+        match self.send(Request::Start).await {
             Ok(Response::Ok) => Ok(()),
             Ok(res) => Err(io::Error::new(
                 io::ErrorKind::InvalidData,
@@ -38,10 +43,10 @@ pub trait Client {
         }
     }
 
-    fn get(&self) -> io::Result<Timer> {
+    async fn get(&self) -> io::Result<Timer> {
         info!("sending request to get timer");
 
-        match self.send(Request::Get) {
+        match self.send(Request::Get).await {
             Ok(Response::Timer(timer)) => {
                 trace!("timer: {timer:#?}");
                 Ok(timer)
@@ -54,10 +59,10 @@ pub trait Client {
         }
     }
 
-    fn set(&self, duration: usize) -> io::Result<()> {
+    async fn set(&self, duration: usize) -> io::Result<()> {
         info!("sending request to set timer duration");
 
-        match self.send(Request::Set(duration)) {
+        match self.send(Request::Set(duration)).await {
             Ok(Response::Ok) => Ok(()),
             Ok(res) => Err(io::Error::new(
                 io::ErrorKind::InvalidData,
@@ -67,10 +72,10 @@ pub trait Client {
         }
     }
 
-    fn pause(&self) -> io::Result<()> {
+    async fn pause(&self) -> io::Result<()> {
         info!("sending request to pause timer");
 
-        match self.send(Request::Pause) {
+        match self.send(Request::Pause).await {
             Ok(Response::Ok) => Ok(()),
             Ok(res) => Err(io::Error::new(
                 io::ErrorKind::InvalidData,
@@ -80,10 +85,10 @@ pub trait Client {
         }
     }
 
-    fn resume(&self) -> io::Result<()> {
+    async fn resume(&self) -> io::Result<()> {
         info!("sending request to resume timer");
 
-        match self.send(Request::Resume) {
+        match self.send(Request::Resume).await {
             Ok(Response::Ok) => Ok(()),
             Ok(res) => Err(io::Error::new(
                 io::ErrorKind::InvalidData,
@@ -93,10 +98,10 @@ pub trait Client {
         }
     }
 
-    fn stop(&self) -> io::Result<()> {
+    async fn stop(&self) -> io::Result<()> {
         info!("sending request to stop timer");
 
-        match self.send(Request::Stop) {
+        match self.send(Request::Stop).await {
             Ok(Response::Ok) => Ok(()),
             Ok(res) => Err(io::Error::new(
                 io::ErrorKind::InvalidData,
@@ -113,12 +118,12 @@ pub trait Client {
 /// be seen as a helper: by implementing the [`ClientStream::read`]
 /// and the [`ClientStream::write`] functions, the trait can deduce
 /// how to handle a request.
-pub trait ClientStream<T> {
-    fn read(&self, stream: &T) -> io::Result<Response>;
-    fn write(&self, stream: &mut T, req: Request) -> io::Result<()>;
-
-    fn handle(&self, stream: &mut T, req: Request) -> io::Result<Response> {
-        self.write(stream, req)?;
-        self.read(stream)
+#[async_trait]
+pub trait ClientStream: RequestWriter + ResponseReader {
+    async fn handle(&mut self, req: Request) -> io::Result<Response> {
+        self.write(req).await?;
+        self.read().await
     }
 }
+
+impl<T: RequestWriter + ResponseReader> ClientStream for T {}